move pageheader component to layouts
